Refactor multi-ad loaders to use common base classes
* Introduce `BaseMultiBannerLoader` and `BaseMultiInterstitialAdLoader` to consolidate shared logic between legacy and Next-Gen GMA SDK implementations.
* Consolidate `MultiBannerViewListener` and `MultiInterstitialAdListener` into a common package and remove redundant "NextGen" listener variants.
* Move internal `PrebidAdLoader` implementation and SDK state management into the base classes to reduce code duplication.
* Update unit tests and demo activities to reference the shared listeners and base class properties.
* Refine banner selection logic to properly promote already-loaded lower-priority SDKs if a higher-priority request fails.
