+++ PrebidMobile/PrebidMobile-core/src/main/java/org/prebid/mobile/api/multiloadercommon/MultiBannerViewListener.kt
@@ -1,4 +1,4 @@
-package org.prebid.mobile.api.multiadloader.listeners
+package org.prebid.mobile.api.multiloadercommon
 
 import android.view.View
 import org.prebid.mobile.api.data.SdkType
