+++ PrebidMobile/PrebidMobile-core/src/main/java/org/prebid/mobile/api/multiloadercommon/MultiInterstitialAdListener.kt
@@ -1,4 +1,4 @@
-package org.prebid.mobile.api.multiadloader.listeners
+package org.prebid.mobile.api.multiloadercommon
 
 import org.prebid.mobile.api.data.SdkType
