--- PrebidMobile/PrebidMobile-core/src/main/java/org/prebid/mobile/api/multiloadernextgen/listeners/MultiBannerViewListenerNextGen.kt
@@ -1,17 +1,0 @@
-package org.prebid.mobile.api.multiloadernextgen.listeners
-
-import android.view.View
-import org.prebid.mobile.api.data.SdkType
-import org.prebid.mobile.api.rendering.BannerView
-
-interface MultiBannerViewListenerNextGen {
-
-    fun onAdLoaded(adView: View, sdk: SdkType)
-    fun onAdFailed(bannerView: BannerView?, error: String?, sdk: SdkType?)
-    fun onAdClicked(bannerView: BannerView?, sdk: SdkType)
-    fun onImpression(sdk: SdkType)
-    fun onAdClosed(bannerView: BannerView?, sdk: SdkType)
-    fun onAdDisplayed(bannerView: BannerView?, sdk: SdkType)
-    fun onAdOpened(sdk: SdkType)
-
-}
