--- PrebidMobile/PrebidMobile-core/src/main/java/org/prebid/mobile/api/multiloadernextgen/listeners/MultiInterstitialAdListenerNextGen.kt
@@ -1,14 +1,0 @@
-package org.prebid.mobile.api.multiloadernextgen.listeners
-
-import org.prebid.mobile.api.data.SdkType
-
-interface MultiInterstitialAdListenerNextGen {
-
-    fun onAdLoaded(sdk: SdkType)
-    fun onAdFailed(error: String?, sdk: SdkType?)
-    fun onAdDisplayed(sdk: SdkType)
-    fun onAdFailedToShow(error: String?, sdk: SdkType?)
-    fun onAdClicked(sdk: SdkType)
-    fun onAdClosed(sdk: SdkType)
-
-}
